# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "EducationIn": {
        "properties": {
          "degree_raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Degree Raw"
          },
          "end_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Year"
          },
          "field_of_study": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Of Study"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "institution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Institution"
          },
          "start_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Year"
          }
        },
        "title": "EducationIn",
        "type": "object"
      },
      "EducationOut": {
        "properties": {
          "degree_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Degree Level"
          },
          "degree_raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Degree Raw"
          },
          "end_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Year"
          },
          "evidence_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence End"
          },
          "evidence_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Start"
          },
          "field_of_study": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Of Study"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "institution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Institution"
          },
          "origin": {
            "title": "Origin",
            "type": "string"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "start_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Year"
          }
        },
        "required": [
          "id",
          "institution",
          "degree_raw",
          "degree_level",
          "field_of_study",
          "start_year",
          "end_year",
          "evidence_start",
          "evidence_end",
          "position",
          "origin"
        ],
        "title": "EducationOut",
        "type": "object"
      },
...
      "ExperienceIn": {
        "properties": {
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "end_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "is_current": {
            "default": false,
            "title": "Is Current",
            "type": "boolean"
          },
          "start_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "title_raw": {
            "title": "Title Raw",
            "type": "string"
          }
        },
        "required": [
          "title_raw"
        ],
        "title": "ExperienceIn",
        "type": "object"
      },
      "ExperienceOut": {
        "properties": {
          "canonical_title_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Title Id"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "end_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "evidence_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence End"
          },
          "evidence_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Start"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_current": {
            "title": "Is Current",
            "type": "boolean"
          },
          "level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "origin": {
            "title": "Origin",
            "type": "string"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "start_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "title_raw": {
            "title": "Title Raw",
            "type": "string"
          }
        },
        "required": [
          "id",
          "company",
          "title_raw",
          "canonical_title_id",
          "level",
          "start_date",
          "end_date",
          "is_current",
          "summary",
          "evidence_start",
          "evidence_end",
          "position",
          "origin"
        ],
        "title": "ExperienceOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobLocationOut": {
        "properties": {
          "is_remote": {
            "title": "Is Remote",
            "type": "boolean"
          },
          "raw_text": {
            "title": "Raw Text",
            "type": "string"
          }
        },
        "required": [
          "raw_text",
          "is_remote"
        ],
        "title": "JobLocationOut",
        "type": "object"
      },
...
      "MatchJobSummary": {
        "description": "Enough of a job to render and link a match card. Deliberately no\ndescription: matches never serve posting text \u2014 the job-detail endpoint\ndoes, under the display-policy gate.",
        "properties": {
          "apply_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apply Url"
          },
          "arrangement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arrangement"
          },
          "canonical_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Title"
          },
          "canonical_title_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Title Slug"
          },
          "company_name": {
            "title": "Company Name",
            "type": "string"
          },
          "experience_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Level"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "locations": {
            "items": {
              "$ref": "#/components/schemas/JobLocationOut"
            },
            "title": "Locations",
            "type": "array"
          },
          "posted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posted At"
          },
          "salary": {
            "$ref": "#/components/schemas/SalaryOut"
          },
          "title_raw": {
            "title": "Title Raw",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title_raw",
          "canonical_title",
          "canonical_title_slug",
          "company_name",
          "locations",
          "arrangement",
          "experience_level",
          "salary",
          "posted_at",
          "apply_url"
        ],
        "title": "MatchJobSummary",
        "type": "object"
      },
      "MatchProfileSummary": {
        "description": "The candidate side of the match, for context and honesty: how many\nskills actually join the taxonomy, which title families were matched\non, and the derived seniority the experience component used.",
        "properties": {
          "experience_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Level"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "highest_degree": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Highest Degree"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "matched_skill_count": {
            "title": "Matched Skill Count",
            "type": "integer"
          },
          "reviewed_at": {
            "format": "date-time",
            "title": "Reviewed At",
            "type": "string"
          },
          "title_families": {
            "items": {
              "type": "string"
            },
            "title": "Title Families",
            "type": "array"
          },
          "unmapped_skill_count": {
            "title": "Unmapped Skill Count",
            "type": "integer"
          },
          "years_experience": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Years Experience"
          }
        },
        "required": [
          "id",
          "full_name",
          "reviewed_at",
          "matched_skill_count",
          "unmapped_skill_count",
          "title_families",
          "experience_level",
          "years_experience",
          "highest_degree"
        ],
        "title": "MatchProfileSummary",
        "type": "object"
      },
...
      "ProfileDetail": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "education": {
            "items": {
              "$ref": "#/components/schemas/EducationOut"
            },
            "title": "Education",
            "type": "array"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "experiences": {
            "items": {
              "$ref": "#/components/schemas/ExperienceOut"
            },
            "title": "Experiences",
            "type": "array"
          },
          "extracted_at": {
            "format": "date-time",
            "title": "Extracted At",
            "type": "string"
          },
          "extractor_version": {
            "title": "Extractor Version",
            "type": "integer"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "resume": {
            "$ref": "#/components/schemas/ResumeOut"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/SkillOut"
            },
            "title": "Skills",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "resume",
          "full_name",
          "email",
          "phone",
          "extractor_version",
          "extracted_at",
          "reviewed_at",
          "created_at",
          "updated_at",
          "skills",
          "experiences",
          "education"
        ],
        "title": "ProfileDetail",
        "type": "object"
      },
...
      "ProfileUpdate": {
        "description": "Full-document replacement. `reviewed` is the document state, not an\nevent: true sets reviewed_at (now, if not already set), false clears\nit \u2014 editing after review naturally un-reviews by sending false.",
        "properties": {
          "education": {
            "items": {
              "$ref": "#/components/schemas/EducationIn"
            },
            "title": "Education",
            "type": "array"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "experiences": {
            "items": {
              "$ref": "#/components/schemas/ExperienceIn"
            },
            "title": "Experiences",
            "type": "array"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "reviewed": {
            "default": false,
            "title": "Reviewed",
            "type": "boolean"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/SkillIn"
            },
            "title": "Skills",
            "type": "array"
          }
        },
        "required": [
          "skills",
          "experiences",
          "education"
        ],
        "title": "ProfileUpdate",
        "type": "object"
      },
...
      "ResumeOut": {
        "properties": {
          "extracted_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extracted Text"
          },
          "file_size": {
            "title": "File Size",
            "type": "integer"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "parser_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parser Name"
          },
          "parser_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parser Version"
          },
          "uploaded_at": {
            "format": "date-time",
            "title": "Uploaded At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "filename",
          "file_size",
          "page_count",
          "parser_name",
          "parser_version",
          "uploaded_at",
          "extracted_text"
        ],
        "title": "ResumeOut",
        "type": "object"
      },
...
      "SavedJobIdsResponse": {
        "description": "The ids the current user has saved \u2014 the search/detail UIs use this to\nrender the Save button's saved state without leaking anything else.",
        "properties": {
          "job_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Job Ids",
            "type": "array"
          }
        },
        "required": [
          "job_ids"
        ],
        "title": "SavedJobIdsResponse",
        "type": "object"
      },
...
      "SavedJobOut": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/MatchJobSummary"
          },
          "match": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SavedJobMatch"
              },
              {
                "type": "null"
              }
            ]
          },
          "saved_at": {
            "format": "date-time",
            "title": "Saved At",
            "type": "string"
          }
        },
        "required": [
          "job",
          "saved_at",
          "match"
        ],
        "title": "SavedJobOut",
        "type": "object"
      },
      "SavedJobRequest": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "integer"
          }
        },
        "required": [
          "job_id"
        ],
        "title": "SavedJobRequest",
        "type": "object"
      },
      "SavedJobsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SavedJobOut"
            },
            "title": "Items",
            "type": "array"
          },
          "profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MatchProfileSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "profile_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id"
          }
        },
        "required": [
          "profile",
          "profile_id",
          "items"
        ],
        "title": "SavedJobsResponse",
        "type": "object"
      },
...
      "SkillIn": {
        "description": "Keep an existing skill (id) or add a manual one (label, optionally\nmapped to a taxonomy slug so it stays matchable).",
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "technology_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technology Slug"
          }
        },
        "title": "SkillIn",
        "type": "object"
      },
      "SkillOut": {
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "evidence_snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Snippet"
          },
          "evidence_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Start"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "occurrences": {
            "title": "Occurrences",
            "type": "integer"
          },
          "origin": {
            "title": "Origin",
            "type": "string"
          },
          "technology_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technology Id"
          }
        },
        "required": [
          "id",
          "technology_id",
          "label",
          "confidence",
          "occurrences",
          "evidence_snippet",
          "evidence_start",
          "origin"
        ],
        "title": "SkillOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Xcelsior API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/profiles/{profile_id}": {
      "delete": {
        "description": "Delete the profile and its uploaded resume (user-owned data).",
        "operationId": "remove_profile_api_v1_profiles__profile_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Profile",
        "tags": [
          "profiles"
        ]
      },
      "get": {
        "operationId": "get_profile_api_v1_profiles__profile_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Profile",
        "tags": [
          "profiles"
        ]
      },
      "put": {
        "description": "Full-document reconcile: id = keep/update, no id = create as\nmanual, absent = delete. One transaction \u2014 the review screen's Save.",
        "operationId": "put_profile_api_v1_profiles__profile_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Profile",
        "tags": [
          "profiles"
        ]
      }
    },
...
    "/api/v1/saved-jobs": {
      "get": {
        "description": "The saved-jobs dashboard: each saved job with a live match score,\nmatched/missing skills, and reasons drawn from the user's most recently\nreviewed profile.",
        "operationId": "get_saved_jobs_api_v1_saved_jobs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedJobsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Saved Jobs",
        "tags": [
          "saved-jobs"
        ]
      },
      "post": {
        "description": "Save a job. Idempotent; returns the updated saved-id set.",
        "operationId": "create_saved_job_api_v1_saved_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedJobIdsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Saved Job",
        "tags": [
          "saved-jobs"
        ]
      }
    },
...
    "/api/v1/saved-jobs/{job_id}": {
      "delete": {
        "description": "Unsave a job. Idempotent.",
        "operationId": "delete_saved_job_api_v1_saved_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Saved Job",
        "tags": [
          "saved-jobs"
        ]
      }
    },
...
  }
}